pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.26.0",
    "numpy>=2.3.1",
    "openai>=1.97.1",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "requests>=2.32.4",
    "streamlit>=1.47.0",
]
